Reorganize Credits class
