--- src/main/java/legend/game/credits/CreditHeader08.java
@@ -1,13 +1,0 @@
-package legend.game.credits;
-
-import legend.game.credits.Credits.CreditsType;
-
-public class CreditHeader08 {
-  public final int creditIndex_00;
-  public final CreditsType type_04;
-
-  public CreditHeader08(final int creditIndex, final Credits.CreditsType type) {
-    this.creditIndex_00 = creditIndex;
-    this.type_04 = type;
-  }
-}
